+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ History
 Version 2.2.2 (in development)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Sync with upstream pathlib from CPython 3.6.1 (bpo-25803, bpo-29694,
+  bpo-30177).
+
 Version 2.2.1
 ^^^^^^^^^^^^^